pyproject.toml
[project]
name = "main"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "alembic==1.14.0",
    "asyncpg>=0.31.0",
    "fastapi>=0.136.0",
    "httpx>=0.28.1",
    "langchain>=1.2.15",
    "langchain-community>=0.4.1",
    "langchain-google-genai>=4.2.2",
    "langchain-huggingface>=1.2.2",
    "langchain-qdrant>=1.1.0",
    "paramiko>=4.0.0",
    "passlib[bcrypt]==1.7.4",
    "pydantic[email]>=2.12.0",
    "python-dotenv>=1.2.2",
    "python-jose[cryptography]==3.3.0",
    "python-multipart==0.0.20",
    "qdrant-client>=1.17.1",
    "rank-bm25>=0.2.2",
    "scp>=0.15.0",
    "sentence-transformers>=5.4.1",
    "sqlalchemy[asyncio]==2.0.36",
    "uvicorn>=0.44.0",
]
